move ImageNetDataReader and ImageClassificationEvaluator to tensorrt resnet example
