+++ onnxruntime/python/tools/quantization/__init__.py
@@ -3,5 +3,5 @@ from .quantize import QuantizationMode
 from .calibrate import CalibrationDataReader, calculate_calibration_data, get_calibrator, generate_calibration_table
 from .calibrate import calibrate
 from .quant_utils import QuantType, write_calibration_table
-from .evaluate import YoloV3Evaluator, YoloV3VisionEvaluator, ImageClassificationEvaluator
-from .data_reader import YoloV3DataReader, YoloV3VisionDataReader, ImageNetDataReader
+from .evaluate import YoloV3Evaluator, YoloV3VisionEvaluator
+from .data_reader import YoloV3DataReader, YoloV3VisionDataReader
